Separate seeds into files
