+++ scripts/seed.ts
@@ -1,7 +1,7 @@
-import { setupDatabase } from "../cypress/support/db";
+import { seed } from "../cypress/support/db";
 
 async function main() {
-  await setupDatabase();
+  await seed();
 }
 
 main();
